requirements.txt
# Telegram
python-telegram-bot>=20.0

# Async HTTP & HTML parsing
aiohttp>=3.9.0
beautifulsoup4>=4.12.0

# Database
aiosqlite>=0.19.0

# Config
python-dotenv>=1.0.0

# Scheduling
apscheduler>=3.10.0
